- Added possibility for consumers to *not* affect the presentation time (adjustments in output.cpp made based on the buffer_depths of the consumers) of other consumers. This is good when consumers that are usually spontaneously added and removed during playout (image_consumer, ffmpeg_consumer, streaming_consumer and channel_consumer) are used, because it prevents the lag previously experienced when adding/removing such consumers. - Fixed streaming_consumer sometimes crashing the server during destruction.
Former-commit-id: 1c23b5cfcd2cc20b8563c9eea3ae706e749590dd

+++ modules/decklink/util/util.h
@@ -359,7 +359,7 @@ struct configuration
 	{
 	}
 	
-	size_t buffer_depth() const
+	int buffer_depth() const
 	{
 		return base_buffer_depth + (latency == low_latency ? 0 : 1) + (embedded_audio ? 1 : 0);
 	}
